requirements.txt
accelerate==0.18.0
alabaster==0.7.8
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==22.2.0
Babel==2.6.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.2
bleach==6.0.0
blessings==1.6
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
comm==0.1.3
command-not-found==0.3
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
diffusers==0.16.0.dev0
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.3
feedparser==5.2.1
filelock==3.11.0
fonttools==4.39.3
future==0.18.2
gitsome==0.8.0
gpustat==0.6.0
httplib2==0.14.0
huggingface-hub==0.13.4
idna==2.8
imagesize==1.2.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==2.10.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==8.1.0
jupyter-console==6.6.3
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-server==2.5.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
monotonic==1.5
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.5.4
notebook-shim==0.2.2
numpy==1.23.5
numpydoc==0.7.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
packaging==23.1
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.2.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-json-logger==2.0.7
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2023.3.23
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==2.0.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.8.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
soupsieve==2.4
Sphinx==1.8.5
ssh-import-id==5.10
stack-data==0.6.2
systemd-python==234
terminado==0.17.1
tinycss2==1.2.1
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.4.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
xkit==0.0.0
xonsh==0.9.13
zipp==3.15.0
